use maven Multi-Module Management
