--- client/src/main/java/com/github/yeecode/objectlogger/client/ObjectLoggerClientApplication.java
@@ -1,13 +1,0 @@
-package com.github.yeecode.objectlogger.client;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ObjectLoggerClientApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ObjectLoggerClientApplication.class, args);
-	}
-
-}
