+++ client/src/main/java/com/github/yeecode/objectlogger/client/task/LogAttributesTask.java
@@ -5,7 +5,6 @@ import com.github.yeecode.objectlogger.client.http.HttpUtil;
 import com.github.yeecode.objectlogger.client.model.BaseAttributeModel;
 import com.github.yeecode.objectlogger.client.model.OperationModel;
 import com.google.gson.Gson;
-import org.springframework.util.CollectionUtils;
 
 import java.util.Date;
 import java.util.List;
@@ -43,8 +42,7 @@ public class LogAttributesTask implements Runnable {
         try {
             OperationModel operationModel = new OperationModel(objectLoggerConfig.getBusinessAppName(), objectName, objectId, operator,
                     operationName, operationAlias, extraWords, comment, new Date());
-
-            if (!CollectionUtils.isEmpty(baseAttributeModelList)) {
+            if (!baseAttributeModelList.isEmpty()) {
                 operationModel.addBaseActionItemModelList(baseAttributeModelList);
             }
             httpUtil.sendLog(new Gson().toJson(operationModel));
