+++ demo/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+server.port=8081
+yeecode.objectLogger.serverAddress=http://127.0.0.1:8080
+yeecode.objectLogger.businessAppName=ObjectLoggerDemo
+yeecode.objectLogger.autoLogAttributes=true
