+++ server/src/main/resources/application.properties
@@ -1,18 +1,15 @@
 # Base Conf
 server.servlet.context-path=/ObjectLogger
 server.port=8080
-
 # Database Conf
 spring.datasource.driver-class-name=com.mysql.jdbc.Driver
 spring.datasource.url=jdbc:{db}://{db_address}/{db_name}
 spring.datasource.username={db_username}
 spring.datasource.password={db_password}
-
 # ObjectLogger Conf
 yeecode.objectLogger.serverAddress=http://127.0.0.1:8080
 yeecode.objectLogger.businessAppName=ObjectLogger
 yeecode.objectLogger.autoLogAttributes=true
-
 # Mybatis Conf
 mybatis.config-location=classpath:mybatis/mybatis-config.xml
 mybatis.mapper-locations=classpath:mybatis/mapper/*.xml
