--- server/src/test/java/com/github/yeecode/objectLogger/server/ObjectLoggerApplicationTests.java
@@ -1,16 +1,0 @@
-package com.github.yeecode.objectLogger.server;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ObjectLoggerApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
