+++ server/src/test/java/com/github/yeecode/objectLogger/server/ObjectLoggerServerApplicationTests.java
@@ -1,0 +1,16 @@
+package com.github.yeecode.objectLogger.server;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class ObjectLoggerServerApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
